+++ CHANGELOG.md
@@ -16,6 +16,12 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 - The list command now has grid printing.
 - The regex search, to search packages based on a given regex pattern.
 - Add support for applying patches to the source code of packages.
+- Verifier and Repairer features
+    - The check and fix command now except multiple package parameters.
+    - Add support for 'check dependencies' (checks which have to be executed before the current check).
+    - The inconsistent register fix doesn't require a re-install anymore.
+    - Add check after fix (re-run check to make sure the fix worked).
+    - Add checks and fixes for: stray directories, empty directories, package existence, correct permissions, missing Config.toml and missing Installed.toml.
 
 ### Changes
 - The build system now includes a new `build-install` xtask to create a full Packit build in a `bin` directory structure.